requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.2
mysql-connector-python==9.0.0
new-package==0.0.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
